requirements.txt
asgiref==3.2.7
astroid==2.4.1
Click==7.0
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
pylint==2.5.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.1
toml==0.10.1
torch==1.3.1
torchvision==0.4.2
typed-ast==1.4.1
virtualenv==16.7.7
Werkzeug==0.16.0
wrapt==1.12.1
